pyproject.toml
[project]
name = "jupyter-notebook-server"
version = "0.1.0"
description = "Jupyter notebook server"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.14.1",
    "apscheduler>=3.11.0",
    "bs4>=0.0.2",
    "build>=1.2.2.post1",
    "certifi>=2025.1.31",
    "cryptography>=44.0.3",
    "fastapi>=0.115.8",
    "ipykernel>=6.29.5",
    "jinja2>=3.1.5",
    "nest-asyncio>=1.6.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pip-system-certs>=4.0",
    "playwright>=1.50.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "python-telegram-bot[job-queue]==21.10",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.38",
    "twilio>=9.4.5",
    "urllib3==1.26.12",
    "uvicorn>=0.34.0",
    "xlsx2pdf>=1.0.4",
]

[tool.setuptools]
packages = [
    "server",
    "database",
    "nepse",
    "services",
    "utils"
]
